Implement AUTO_PARTITION_MODE
